requirements.txt
chromadb
fastapi
httpx
keybert
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
openai
pandas
python-dotenv
python-multipart
scikit-learn
sentence-transformers
supabase
tiktoken
torch
transformers
uvicorn
